Implement a generalized entity script and scene
This should allow for the creation of multiple
entities in the game. To add an entity, create
an inherited scene for the entity and modify the
sprite and hitbox nodes. Then, add instances of
the entity to the main scene (don't forget to
add the new entities to the `entities` group)